[QOLDEV-1284]Migrated tests to Cypress 15
