+++ .gitignore
@@ -67,6 +67,9 @@ web_modules/
 .yarn/install-state.gz
 .pnp.*
 
+# Cypress generated assets
+cypress/screenshots/
+
 # MacOS
 .DS_Store
 .vscode
